+++ clients/include/blas3/testing_trsm_strided_batched.hpp
@@ -449,7 +449,7 @@ void testing_trsm_strided_batched(const Arguments& arg)
         make_unit_diagonal(uplo, hA);
     }
 
-    for(int b = 0; b < batch_count; b++)
+    for(int64_t b = 0; b < batch_count; b++)
     {
         // Calculate hB = hA*hX;
         ref_trmm<T>(side,
